+++ src/renderer/modules/common/i18n.ts
@@ -1,7 +1,7 @@
 import { filters, waitForModule } from "../webpack";
 import EventEmitter from "events";
 
-import type { RawModule } from "../../../../types";
+import type { RawModule } from "../../../types";
 import type { Primitive } from "type-fest";
 import type { Rule } from "./parser";
