+++ src/renderer/modules/common/messages.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { filters, waitForModule } from "../webpack";
 
-import type { RawModule } from "../../../../types";
+import type { RawModule } from "../../../types";
 import type { Message, MessageAttachment } from "discord-types/general";
 
 export enum ActivityActionTypes {
